Make the test program encode audio in the Ogg Vorbis format.
The test program is not trying to produce lossless rips, but rather
easy-to-modify rips that discard as much useless data as possible.

FLAC is lossless, but it preserves a lot of useless data that cannot
be heard anyway and barely saves any space.

Opus is supposedly a better lossy format, but Mega CD emulators are
unlikely to support it and it can only encode that is 48kHz, not
44.1kHz.

MP3 has been found to be a worse lossy format than Ogg Vorbis in
terms of audio quality.